Extracted parts of Yarp resource into abstract WebApplication resource
